requirements.txt
python-telegram-bot[webhooks]==20.7
reportlab==4.1.0
python-dotenv==1.0.0
